requirements.txt
torch
whisper
pyannote.audio
scikit-learn
numpy
langchain-ollama
ffmpeg-python
fastapi
uvicorn